fix content of /proc/sys/crypto/fips_enabled

+++ vhdbuilder/packer/post-install-dependencies.sh
@@ -16,17 +16,19 @@ VHD_LOGS_FILEPATH=/opt/azure/vhd-install.complete
 MAX_BLOCK_COUNT=30298176 # 30 GB
 
 if [[ $OS == $UBUNTU_OS_NAME ]]; then
-  # shellcheck disable=SC2021
-  current_kernel="$(uname -r | cut -d- -f-2)"
-  dpkg --get-selections | grep -e "linux-\(headers\|modules\|image\)" | grep -v "$current_kernel" | tr -s '[[:space:]]' | tr '\t' ' ' | cut -d' ' -f1 | xargs -I{} apt-get remove -yq {}
-
   # remove apport
   apt-get purge --auto-remove apport open-vm-tools -y
 
-  # strip old kernels/packages
-  apt-get -y autoclean || exit 1
-  apt-get -y autoremove --purge || exit 1
-  apt-get -y clean || exit 1
+  if [[ "${ENABLE_FIPS,,}" != "true" ]]; then
+    # shellcheck disable=SC2021
+    current_kernel="$(uname -r | cut -d- -f-2)"
+    dpkg --get-selections | grep -e "linux-\(headers\|modules\|image\)" | grep -v "$current_kernel" | tr -s '[[:space:]]' | tr '\t' ' ' | cut -d' ' -f1 | xargs -I{} apt-get remove -yq {}
+
+    # strip old kernels/packages
+    apt-get -y autoclean || exit 1
+    apt-get -y autoremove --purge || exit 1
+    apt-get -y clean || exit 1
+  fi
 fi
 
 # shellcheck disable=SC2129
